feat: start to do tests
