+++ web-react/tests/setup.tsx
@@ -1,0 +1,19 @@
+import { cleanup, render } from '@testing-library/react';
+import { afterEach } from 'vitest';
+import '@testing-library/jest-dom';
+
+afterEach(() => {
+  cleanup();
+});
+
+function customRender(ui: React.ReactElement, options = {}) {
+  return render(ui, {
+    wrapper: ({ children }) => children,
+    ...options,
+  });
+}
+
+export * from '@testing-library/react';
+export { default as userEvent } from '@testing-library/user-event';
+
+export { customRender as render };
